requirements.txt
moviepy==2.1.2
diffusers
torch
transformers
pysrt
openai-whisper
soundfile
kokoro
accelerate
urllib3<2.0.0
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
requests
pillow
groq
schedule
python-dotenv